fix build errors

+++ src/mmchecking_list.cpp
@@ -732,7 +732,7 @@ void TransactionListCtrl::onMouseRightClick(wxMouseEvent& event)
         );
     }
     bool columnIsAmount = false;
-    int col_nr = static_cast<int>(getColumnFromPosition(event.GetX()));
+    int col_nr = getColumnFromPosition(event.GetX());
     int flags;
     unsigned long row = HitTest(event.GetPosition(), flags);
     if (row < m_trans.size() && (flags & wxLIST_HITTEST_ONITEM) && col_nr < getColNrSize()) {
